requirements.txt
amqp==5.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
bleach==4.1.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.3
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-rosetta==0.9.8
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
filelock==3.1.0
idna==3.4
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==8.1.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
kombu==5.2.4
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mistune==0.8.4
MouseInfo==0.1.3
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.4.0
polib==1.1.1
prometheus-client==0.13.1
prompt-toolkit==3.0.30
pure-eval==0.2.2
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.11.2
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.1
pywin32==303
pywinpty==0.5.7
pyzmq==24.0.1
redis==4.3.4
requests==2.28.1
Send2Trash==1.8.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
terminado==0.9.5
testpath==0.6.0
tornado==6.1
traitlets==5.1.1
tzdata==2022.2
urllib3==1.26.12
vine==5.0.0
virtualenv==20.8.1
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
zipp==3.7.0
